\name{FIQ400}
\alias{FIQ400}
\docType{data}
\title{
FIQ database.
}
\description{
A database to be used as example in the functions included on \code{SINRELEF.LD} package. It is part of the data from the study by Dueñas et al. (2022).
}
\usage{data("FIQ400")}
\format{
  A data frame with 400 observations and 17 variables, ranged between 1 and 4.
}
\details{
The participants in this study were parents of adolescents who were enrolled in secondary school. The parents completed the Spanish adaptation of the Family Involvement Questionnaire-High School Version (FIQ-HS), which assesses parental family involvement in the education of their sons and daughters. More specifically, we have randomly selected 400 participants from this study, and we have used their data in the 17 items of the Home-based activities subscale. These items have 4 response options (rarely, sometimes, often and always). We have chosen this subscale because the study by Dueñas et a. (2022) showed that the error terms of some pairs of items were substantially correlated. According to the authors, this was expected because the corresponding item stems were either very similarly worded or tapped similar (although not identical) content.
}
\references{
Dueñas, J. M., Morales-Vives, F., Camarero-Figuerola, M., Tierno-García, J. M. (2022). Spanish adaptation of The Family Involvement Questionnaire-High School: Version for parents. Psicología Educativa, 28(1), 31-38. \doi{10.5093/psed2020a21}
}
\examples{
data(FIQ400)
}
\keyword{datasets}
