\name{SINRELEF.LD-internal}
\alias{SINRELEF.LD-internal}
\alias{bivnor}
\alias{contiteor}
\alias{discreB}
\alias{doubletrelefconti}
\alias{doubletrelefcontim}
\alias{doubletrelefuva}
\alias{erf}
\alias{especov}
\alias{espeit}
\alias{fnor3}
\alias{gauss}
\alias{normalbi}
\alias{omegaldconti}
\alias{omegalduva}
\alias{ordnor}
\alias{pmpolycho}
\alias{ppolyser}
\alias{relefdeleted}
\alias{relefdeleteduva}
\alias{replicbase}
\alias{replicdiscre}
\alias{rreplicbase}
\alias{rreplicdiscre}
\alias{size}
\alias{transpose}
\alias{FIQ_CONT_ITEMS_D}
\alias{FIQ_CONT_L}
\alias{FIQ_CONT_VALUES_D}
\alias{FIQ_CONT_VAR}
\alias{.Random.seed}
\title{
SINRELEF-internal
}
\description{
This functions are for internal use only.
}
\usage{
bivnor
contiteor
discreB
doubletrelefconti
doubletrelefcontim
doubletrelefuva
erf
especov
espeit
fnor3
gauss
normalbi
omegaldconti
omegalduva
ordnor
pmpolycho
ppolyser
relefdeleted
relefdeleteduva
replicbase
replicdiscre
rreplicbase
rreplicdiscre
size
transpose
FIQ_CONT_ITEMS_D
FIQ_CONT_L
FIQ_CONT_VALUES_D
FIQ_CONT_VAR
}
\value{
No return value for the user.
}
\author{
David Navarro-Gonzalez

Fabia Morales-Vives

Pere Joan Ferrando
}
\keyword{ internal }
