\name{SINRELEF.LD_GUI}
\alias{SINRELEF.LD_GUI}

\title{
Score Information, reliability, and relative efficiency in measures that contain locally-dependent items, using a Graphical User Interface.
}
\description{
A Graphical User Interface (GUI) caller for the \code{\link{SINRELEF.LD}} main function. The GUI contains all the required inputs for the execution of \code{\link{SINRELEF.LD}}, as well as the output display on console.

For more info about the inputs required, or the output information, please \code{\link{SINRELEF.LD}}.
}
\usage{
SINRELEF.LD_GUI()
}
\author{
David Navarro-Gonzalez

Fabia Morales-Vives

Pere Joan Ferrando
}
\examples{

#Type SINRELEF.LD_GUI() to open the GUI.

}
